Added slide support to robot class, plus some more
Added in slide motor definitions to the robot class, also added the ability to turn on the brakes.

I edited NewManualOp to use the new slide definitions and drive motor braking.

Changed some of the headless code in HeadlessOp, comments explain my thought process more.

DO NOT use any Opmode that uses the Robot class until I am there to supervise, it is completely untested and I need to be there in case something breaks. This includes HeadlessOp and NewManualOp.