Created proof of concepts for autonomous funcions
Added a target region to the parkingZoneFinder pipeline. This will need to be changed to properly fit the signal cone as we go.

Created a proof of concept opmode to display the parking zone during init

Created a proof of concept to interrupt an asynchronous trajectory. This is important for my future plans for angle adjustment